requirements.txt
alembic==1.6.2
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.0
dominate==2.6.0
Flask==2.0.0
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
urllib3==1.26.5
uWSGI==2.0.17
visitor==0.1.3
Werkzeug==2.0.0
